python-ai/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
torch==2.1.2
torchvision==0.16.2
ultralytics==8.0.215
opencv-python==4.9.0.80
Pillow==10.1.0
numpy==1.24.3
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
boto3==1.34.7
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
tenacity==8.2.3
pydantic==2.5.0
python-dotenv==1.0.0